requirements.txt
flask==3.0.3
flask-pymongo==2.3.0
pymongo[srv]==4.7.2
bcrypt==4.1.3
PyJWT==2.8.0
python-dotenv==1.0.1
dnspython==2.6.1
gunicorn==22.0.0
